# web6/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FundedFrens API — Solana meme coin prop trading platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/challenge-plans/{slug}:
    get:
      operationId: getChallengePlan
      tags: [plans]
      summary: Get a challenge plan by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengePlan"
...
  /v1/orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order and generate SOL payment details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "200":
          description: Created order with payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List my orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
  /v1/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a specific order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /v1/orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /v1/orders/{orderId}/payment-status:
    get:
      operationId: pollPaymentStatus
      tags: [orders]
      summary: Poll payment status — call every 5 seconds from payment page
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"
  /v1/orders/{orderId}/confirm:
    post:
      operationId: confirmOrder
      tags: [orders]
      summary: Confirm payment with a Solana transaction signature — activates the challenge
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderConfirmInput"
      responses:
        "200":
          description: Activated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ChallengePlan:
      type: object
      required: [id, name, slug, price_usd, funded_sol, status, display_order, profit_target_pct, max_drawdown_pct, daily_drawdown_pct, min_trading_days, max_position_size_pct, max_open_positions, reactivation_cost_pct]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price_usd:
          type: number
        funded_sol:
          type: number
        funded_usd_estimate:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, inactive, coming_soon, archived]
        display_order:
          type: integer
        description:
          type: ["string", "null"]
        profit_target_pct:
          type: number
        max_drawdown_pct:
          type: number
        daily_drawdown_pct:
          type: number
        min_trading_days:
          type: integer
        max_position_size_pct:
          type: number
        max_open_positions:
          type: integer
        reactivation_cost_pct:
          type: number
        created_at:
          type: ["string", "null"]
        updated_at:
          type: ["string", "null"]
...
    OrderInput:
      type: object
      required: [challenge_plan_id]
      properties:
        challenge_plan_id:
          type: integer
        referral_code:
          type: ["string", "null"]
    OrderConfirmInput:
      type: object
      required: [tx_signature]
      properties:
        tx_signature:
          type: string
    Order:
      type: object
      required: [id, user_id, status, amount, currency, created_at, updated_at]
      properties:
        id:
          type: integer
        user_id:
          type: string
        challenge_plan_id:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, waiting_for_payment, payment_detected, confirming, confirmed, activated, expired, cancelled, failed]
        amount:
          type: number
        currency:
          type: string
        plan_name:
          type: ["string", "null"]
        plan_slug:
          type: ["string", "null"]
        expected_sol:
          type: ["number", "null"]
        sol_price_usd:
          type: ["number", "null"]
        treasury_wallet:
          type: ["string", "null"]
        expiry_time:
          type: ["string", "null"]
        payment_status:
          type: ["string", "null"]
        tx_signature:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    OrderList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PaymentStatus:
      type: object
      required: [order_id, order_status, challenge_activated]
      properties:
        order_id:
          type: integer
        order_status:
          type: string
        payment_status:
          type: ["string", "null"]
        expected_sol:
          type: ["number", "null"]
        received_sol:
          type: ["number", "null"]
        treasury_wallet:
          type: ["string", "null"]
        tx_signature:
          type: ["string", "null"]
        expiry_time:
          type: ["string", "null"]
        confirmed_at:
          type: ["string", "null"]
        seconds_remaining:
          type: ["integer", "null"]
        challenge_activated:
          type: boolean
...